Eliminate Array Copy
Try to eliminate array copying, when
truncation pattern is deteced, in particularly, when:
* array is not used after copying;
* destination array is smaller than source one;
* copy size is equal to destination size;
* source copy offset and destination offset both are zero
Such pattern is commonly visible in StringBuffer, ByteArrayOutputStream.

When truncation pattern is detected the ideal graph is reshaped
to conditionally skip copying and to use source array in place
of destination one.

The souce's length is adjusted and dummy array is put in source tail
to fill gap in heap, which if existed would make garbage
collection fail; changes are performed in reverse order
so GC would not see new size, before tail space is filled up.

The source array have to be big enough to carray at least empty
array with proper oop header. This checks are performed at runtime
as during compilation sizes, nor even destination length and
copy length are not known (see Arrays.copyOf).

Becouse the new header is put in tail sapce, this is limitted only
to primitive arrays (when scanning GC will read header bytes not
pointing to oop), but can be extend after small adjustments to object
arrays.

The usage checkes requires more work as array can be used after
copy, when it's set to fields of non-esacaping objects,
it a part of multi-dimensional array and so on. Simillary, escape
analyse should not remove locks and mark source as not escaping,
as due to conditioanl skip source can escape as destination.

Signed-off-by: Radek Smogura <radek@smogura.eu>